split full test page components
